style(sha512): comply to the norm
