+++ include/sha256_algo.h
@@ -6,7 +6,7 @@
 /*   By: bdevessi <baptiste@devessier.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/24 15:34:55 by bdevessi          #+#    #+#             */
-/*   Updated: 2021/01/05 16:55:59 by bdevessi         ###   ########.fr       */
+/*   Updated: 2021/01/07 23:57:22 by bdevessi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,11 +25,11 @@ void		fill_sha256_states(t_sha256_states *to_fill
 void		add_sha256_states(uint32_t intermediate_hash[8]
 	, t_sha256_states to_add);
 
-uint32_t	ch(uint32_t x, uint32_t y, uint32_t z);
-uint32_t	maj(uint32_t x, uint32_t y, uint32_t z);
-uint32_t	bsig0(uint32_t x);
-uint32_t	bsig1(uint32_t x);
-uint32_t	ssig0(uint32_t x);
-uint32_t	ssig1(uint32_t x);
+uint32_t	ch_32(uint32_t x, uint32_t y, uint32_t z);
+uint32_t	maj_32(uint32_t x, uint32_t y, uint32_t z);
+uint32_t	bsig0_32(uint32_t x);
+uint32_t	bsig1_32(uint32_t x);
+uint32_t	ssig0_32(uint32_t x);
+uint32_t	ssig1_32(uint32_t x);
 
 #endif
