+++ src.mk
@@ -22,6 +22,9 @@ SRCS += srcs/sha256_algo_utils.c
 SRCS += srcs/sha256_print.c
 SRCS += srcs/sha512.c
 SRCS += srcs/sha512_algo.c
+SRCS += srcs/sha512_algo_logical_functions.c
+SRCS += srcs/sha512_algo_logical_functions_2.c
+SRCS += srcs/sha512_algo_utils.c
 SRCS += srcs/sha512_print.c
 SRCS += srcs/ssl.c
 SRCS += srcs/usage.c
